moving rangeDisplayText to helpers
